Delete experiment module and mrta_dataset dependency
Before, datasets were imported from the mrta_dataset repository. The ccu and robot were expecting a dataset module and a file.
This commit re-adds the test module with an allocation test and a dataset. The experiment implementation will be re-designed and added in a future commit
