+++ mrs/robot.py
@@ -4,12 +4,11 @@ import time
 
 from mrs.config.builders import robot
 from fmlib.config.params import ConfigParams
-from mrs.utils.datasets import validate_dataset_file
 
 ConfigParams.default_config_module = 'mrs.config.default'
 
 
-def get_robot_components(robot_id, experiment_name, dataset_file, config_file=None):
+def get_robot_components(robot_id, config_file=None):
     if config_file is None:
         config_params = ConfigParams.default()
     else:
@@ -18,22 +17,9 @@ def get_robot_components(robot_id, experiment_name, dataset_file, config_file=No
     logger_config = config_params.get('logger')
     logging.config.dictConfig(logger_config)
 
-    experiment_config = get_experiment_config(experiment_name, dataset_file, config_params)
-
-    components = robot.configure(robot_id, config_params, experiment_config=experiment_config)
+    components = robot.configure(robot_id, config_params)
 
     return components
-
-
-def get_experiment_config(experiment_name, dataset_file, config_params):
-    dataset_module, dataset_file = validate_dataset_file(experiment_name, dataset_file)
-    experiment_config = {'experiment_name': experiment_name,
-                         'port': config_params.get('experiment_store').get('port'),
-                         'dataset_module': dataset_module,
-                         'dataset_file': dataset_file,
-                         'new_run': False}
-
-    return experiment_config
 
 
 class Robot(object):
@@ -79,15 +65,9 @@ if __name__ == '__main__':
 
     parser.add_argument('robot_id', type=str, help='example: ropod_001')
 
-    parser.add_argument('experiment_name', type=str, action='store', help='Name of the experiment',
-                        choices=['non_intentional_delays',
-                                 'intentional_delays'])
-
-    parser.add_argument('dataset_file', type=str, action='store', help='Name of the file, including extension')
-
     args = parser.parse_args()
 
-    robot_components = get_robot_components(args.robot_id, args.experiment_name, args.dataset_file, args.file)
+    robot_components = get_robot_components(args.robot_id, args.file)
 
     new_robot = Robot(args.robot_id, **robot_components)
     new_robot.run()
