+++ mrs/tests/data/non_overlapping.yaml
@@ -1,0 +1,68 @@
+dataset_id: 8774f36b-d0d9-469b-8414-90985d6af1c7
+dataset_name: non_overlapping_1
+dataset_type: non_overlapping_tw
+interval_type: random
+lower_bound: 60
+task_type: mrs.structs.allocation
+tasks:
+  002767db-01f3-4e0c-af4e-52c4d81bd713:
+    earliest_start_time: 10
+    estimated_duration: 9.12
+    finish_location: second_door_outside
+    hard_constraints: true
+    task_id: 002767db-01f3-4e0c-af4e-52c4d81bd713
+    latest_start_time: 13
+    start_location: table
+    status:
+      status: 1
+      task_id: 002767db-01f3-4e0c-af4e-52c4d81bd713
+      delayed: false
+  173d27eb-d0a4-4b94-986b-1e9b8eca0579:
+    earliest_start_time: 20
+    estimated_duration: 6.84
+    finish_location: second_door_outside
+    hard_constraints: true
+    task_id: 173d27eb-d0a4-4b94-986b-1e9b8eca0579
+    latest_start_time: 23
+    start_location: inspection_area
+    status:
+      status: 1
+      task_id: 173d27eb-d0a4-4b94-986b-1e9b8eca0579
+      delayed: false
+  4076f78b-16d8-4057-a1ad-d16d239764fa:
+    earliest_start_time: 30
+    estimated_duration: 2.74
+    finish_location: shelf
+    hard_constraints: true
+    task_id: 4076f78b-16d8-4057-a1ad-d16d239764fa
+    latest_start_time: 33
+    start_location: cupboard
+    status:
+      status: 1
+      task_id: 002767db-01f3-4e0c-af4e-52c4d81bd713
+      delayed: false
+  6d17963b-10b5-4143-b279-84d8e93aaaa3:
+    earliest_start_time: 40
+    estimated_duration: 2.3
+    finish_location: observation_pose
+    hard_constraints: true
+    task_id: 6d17963b-10b5-4143-b279-84d8e93aaaa3
+    latest_start_time: 43
+    start_location: living_room
+    status:
+      status: 1
+      task_id: 6d17963b-10b5-4143-b279-84d8e93aaaa3
+      delayed: false
+  73ace543-260e-474a-9112-3abd470ed5b5:
+    earliest_start_time: 50
+    estimated_duration: 8.24
+    finish_location: kitchen
+    hard_constraints: true
+    task_id: 73ace543-260e-474a-9112-3abd470ed5b5
+    latest_start_time: 53
+    start_location: shelf
+    status:
+      status: 1
+      task_id: 73ace543-260e-474a-9112-3abd470ed5b5
+      delayed: false
+upper_bound: 300
