+++ mrs/utils/datasets.py
@@ -4,39 +4,8 @@ from datetime import timedelta
 import yaml
 from fmlib.models.requests import TransportationRequest
 from fmlib.models.tasks import Task
-from fmlib.utils.utils import load_file_from_module
 from ropod.utils.timestamp import TimeStamp
 from ropod.utils.uuid import generate_uuid
-from importlib_resources import contents
-
-
-def get_dataset_module(experiment_name):
-    """ Returns the dataset module for the experiment_name
-    """
-    if experiment_name == 'non_intentional_delays':
-        dataset_module = 'dataset_lib.datasets.non_overlapping_tw.generic_task.random'
-    elif experiment_name == 'intentional_delays':
-        dataset_module = 'dataset_lib.datasets.non_overlapping_tw.generic_task.random'
-
-    return dataset_module
-
-
-def get_dataset_files(dataset_module):
-    dataset_files = list()
-    files = contents(dataset_module)
-    for file in files:
-        if file.endswith('.yaml'):
-            dataset_files.append(file)
-
-    return dataset_files
-
-
-def validate_dataset_file(experiment_name, dataset_file):
-    dataset_module = get_dataset_module(experiment_name)
-    dataset_files = get_dataset_files(dataset_module)
-    if dataset_file not in dataset_files:
-        raise ValueError(dataset_file)
-    return dataset_module, dataset_file
 
 
 def load_yaml(file):
@@ -50,11 +19,10 @@ def load_yaml(file):
     return data
 
 
-def load_tasks_to_db(dataset_module, dataset_file):
-    file = load_file_from_module(dataset_module, dataset_file)
-    dataset = yaml.safe_load(file)
+def load_tasks_to_db(dataset_path):
+    dataset_dict = load_yaml(dataset_path)
 
-    tasks_dict = dataset.get('tasks')
+    tasks_dict = dataset_dict.get('tasks')
     ordered_tasks = collections.OrderedDict(sorted(tasks_dict.items()))
     tasks = list()
 
@@ -75,12 +43,6 @@ def load_tasks_to_db(dataset_module, dataset_file):
         tasks.append(task)
 
     return tasks
-
-
-def get_dataset_id(dataset_module, dataset_file):
-    file = load_file_from_module(dataset_module, dataset_file)
-    dataset = yaml.safe_load(file)
-    return dataset.get('dataset_id')
 
 
 def reference_to_current_time(earliest_time, latest_time):
